requirements.txt
streamlit
openai
ffmpeg-python
openai-whisper
faster-whisper

